fix(mongodb): ignore invalid URLs
